create-folder-blocks-and-share-function
