+++ .stylelintrc.yml
@@ -7,11 +7,15 @@ extends:
 
 plugins: ['stylelint-scss']
 rules:
+    # 'comment-no-empty': null
+    # 'scss/comment-no-empty': true
     # 'at-rule-no-unknown': null
     # 'scss/at-rule-no-unknown': true
     # 'scss/dollar-variable-pattern': '^color'
     # 'scss/selector-no-redundant-nesting-selector': true
     # 'color-function-notation': legacy
     # 'font-family-no-missing-generic-family-keyword': true
+    Comment:
+    enabled: false
 
 overrides: [{ files: ['**/*.scss'], customSyntax: 'postcss-scss' }]
