+++ src/styles/_page-with-cards.scss
@@ -154,8 +154,8 @@
     }
 }
 
-.block-game__card-wrapper_background .card-block__background {
-}
+// .block-game__card-wrapper_background .card-block__background {
+// }
 
 .attempt {
     color: #fff;
