+++ src/ts/blocks/blocks-for-game-page/create-button.ts
@@ -1,0 +1,11 @@
+export function createButton(container: HTMLDivElement) {
+    const button = document.createElement('button');
+    button.classList.add('button', 'block-game__button');
+    button.textContent = 'Начать заново';
+
+    button.addEventListener('click', () => {
+        window.application.renderScreen('render-start-page');
+    });
+
+    container.appendChild(button);
+}
