+++ src/ts/state/init-game-state.ts
@@ -1,21 +1,19 @@
+import { createForm } from '../blocks/create-form';
+import { createTimerForGame } from '../blocks/blocks-for-game-page/create-timer';
+import { createAttempts } from '../blocks/blocks-for-game-page/create-attempts';
+import { createButton } from '../blocks/blocks-for-game-page/create-button';
+import { createCards } from '../blocks/blocks-for-game-page/create-cards';
 import {
-    createForm,
-    createTimerForGame,
-    createButton,
-    createCards,
-    createAttempts,
     createBlockWin,
     createBlockLoss,
-    createBlockresultTime,
-    createButtonAgain,
-} from '../blocks';
+} from '../blocks/block-result-game/create-block-result';
+import { createBlockresultTime } from '../blocks/block-result-game/create-block-result-time';
+import { createButtonAgain } from '../blocks/block-result-game/create-button-again';
 
-import {
-    renderStartPage,
-    renderGamePage,
-    renderWinPage,
-    renderLossPage,
-} from '../screens';
+import { renderStartPage } from '../screens/start-page';
+import { renderGamePage } from '../screens/game-page';
+import { renderWinPage } from '../screens/win-page';
+import { renderLossPage } from '../screens/loss-page';
 
 export function initGameState() {
     return (window.application = {
